Consolidate Node tests
